api/requirements.txt
alembic==1.0.8
aniso8601==5.1.0
certifi==2018.11.29
Click==7.0
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.7
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
relativetimebuilder==0.2.0
six==1.12.0
SQLAlchemy==1.3.0
Werkzeug==0.12.2
hashids==1.2.0
PyMySQL==0.9.3
pyjwt==1.7.1
Sphinx==3.1.1
sphinx-rtd-theme==0.4.3
pyecharts==1.8.1
Flask-Mail==0.9.1
pycryptodome==3.10.1
yagmail==0.14.256
pandas==1.2.4
xlrd==2.0.1
Flask-Limiter==1.4
flask-redis==0.4.0
flasgger==0.9.5
